some clean up
