+++ src/main/java/com/hedvig/backoffice/security/dto/Role.kt
@@ -1,0 +1,9 @@
+package com.hedvig.backoffice.security.dto
+
+enum class Role {
+  NOBODY,
+  IEX,
+  IEX_EXTENDED,
+  DEV,
+  ROOT
+}
