+++ src/main/java/com/hedvig/backoffice/security/dto/TokenInfo.kt
@@ -1,0 +1,7 @@
+package com.hedvig.backoffice.security.dto
+
+data class TokenInfo(
+    val subject: String,
+    val scopes: Set<String>,
+    val role: Role
+)
