+++ src/main/java/com/hedvig/backoffice/services/gatekeeper/GatekeeperService.kt
@@ -1,0 +1,7 @@
+package com.hedvig.backoffice.services.gatekeeper
+
+import com.hedvig.backoffice.security.dto.TokenInfo
+
+interface GatekeeperService {
+  fun tokenInfo(authorization: String): TokenInfo
+}
